Revert "Feature/base object model"
