+++ .gitignore
@@ -53,6 +53,3 @@ media/
 
 .elasticbeanstalk/
 aws_credential_file
-
-# PyCharm
-.idea
