--- create_migration.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-source env/bin/activate
-./manage.py schemamigration rfmedia --auto --settings=reformedacademy.local_settings
-./manage.py schemamigration reformedacademy --auto --settings=reformedacademy.local_settings
