+++ reformedacademy/models.py
@@ -5,57 +5,66 @@ Defines models for the reformedacademy app.
 Created by kabucey
 
 """
+from __future__ import unicode_literals
 from django.db import models
-from django.contrib.auth.models import User
-from rfmedia.models import Asset
+from django.core.urlresolvers import reverse
 
 
-class ActivationKey(models.Model):
-    user = models.ForeignKey(User)
-    key = models.CharField(max_length=255)
-    sent = models.DateTimeField(auto_now=True)
+class Type(models.Model):
+    name = models.CharField(max_length=255, blank=True, null=True)
+    created = models.DateTimeField()
+    modified = models.DateTimeField(blank=True, null=True)
 
-
-class Category(models.Model):
-    name = models.CharField(max_length=255)
-    slug = models.SlugField(max_length=255)
+    class Meta:
+        db_table = 'types'
 
     def __unicode__(self):
         return self.name
 
 
-class Course(models.Model):
-    category = models.ForeignKey(Category)
-    name = models.CharField(max_length=255)
-    slug = models.SlugField(max_length=255)
-    description = models.TextField()
+class Asset(models.Model):
+    type = models.ForeignKey(Type)
+    tag = models.CharField(max_length=255, blank=True, null=True)
+    url = models.CharField(max_length=255, blank=True, null=True)
+    filesize = models.BigIntegerField(blank=True, null=True)
+    duration = models.CharField(max_length=255, blank=True, null=True)
+    mime_type = models.CharField(max_length=255, blank=True, null=True)
+    thumbnail_url = models.CharField(max_length=255, blank=True, null=True)
+    active = models.IntegerField()
+    created = models.DateTimeField()
+    modified = models.DateTimeField(blank=True, null=True)
+
+    @property
+    def download_url(self):
+        return reverse('download_asset', args=['web', self.type.name, '{}.mp3'.format(self.tag)])
+
+    class Meta:
+        db_table = 'assets'
 
     def __unicode__(self):
-        return self.name
+        return self.url
 
 
-class Lesson(models.Model):
-    course = models.ForeignKey(Course)
-    name = models.CharField(max_length=255)
-    slug = models.SlugField(max_length=255)
-    description = models.TextField()
-    order = models.PositiveIntegerField()
+class Stat(models.Model):
+    asset = models.ForeignKey(Asset)
+    ip = models.CharField(max_length=255, blank=True)
+    useragent = models.CharField(max_length=255, blank=True)
+    referer = models.CharField(max_length=255, blank=True)
+    method = models.CharField(max_length=4, blank=True)
+    country_code = models.CharField(max_length=45, blank=True)
+    country_name = models.CharField(max_length=255, blank=True)
+    region_code = models.CharField(max_length=45, blank=True)
+    region_name = models.CharField(max_length=45, blank=True)
+    city = models.CharField(max_length=255, blank=True)
+    zipcode = models.CharField(max_length=45, blank=True)
+    metrocode = models.CharField(max_length=45, blank=True)
+    latitude = models.FloatField(blank=True, null=True)
+    longitude = models.FloatField(blank=True, null=True)
+    created = models.DateTimeField()
+    modified = models.DateTimeField(blank=True, null=True)
+
+    class Meta:
+        db_table = 'stats'
 
     def __unicode__(self):
-        return self.name
-
-
-class Task(models.Model):
-    lesson = models.ForeignKey(Lesson)
-    asset = models.ForeignKey(Asset, on_delete=models.PROTECT)
-    order = models.PositiveIntegerField()
-
-    def __unicode__(self):
-        return '{} {}'.format(self.lesson, self.asset.tag)
-
-
-class Teacher(models.Model):
-    course = models.ManyToManyField(Course)
-    name = models.CharField(max_length=255)
-    biography = models.TextField()
-    profile_image = models.ImageField(upload_to='profile_images/teachers')
+        return '{} {}'.format(self.asset, self.ip)
