--- reformedacademy/templates/reformedacademy/404.html
@@ -1,14 +1,0 @@
-{% extends 'reformedacademy/base_no_menu.html' %}
-{% load staticfiles %}
-
-{% block content %}
-    <div class="top-container">
-        <div class="inner">
-            <img src="{% static 'images/lion.png' %}" alt="Logo">
-            <h1>Uh Oh.</h1>
-            <div class="description">
-                The page you were looking for cannot be found. Please visit our <a href="{% url 'support' %}">support page</a>.
-            </div>
-        </div>
-    </div>
-{% endblock %}
