--- reformedacademy/templates/reformedacademy/account_created.html
@@ -1,15 +1,0 @@
-{% extends 'reformedacademy/base_no_menu.html' %}
-{% load staticfiles %}
-
-{% block content %}
-    <div class="top-container centered activation">
-        <div class="inner">
-            <img src="{% static 'images/lion.png' %}" alt="Logo">
-            <h1>Check Your Email!</h1>
-            <div class="description">
-                Instructions on how to activate your account have been sent to the email address you provided.
-                You'll be automatically directed back to Reformed Academy.<br><br>See you soon!
-            </div>
-        </div>
-    </div>
-{% endblock %}
