--- reformedacademy/templates/reformedacademy/base.html
@@ -1,77 +1,0 @@
-{% load staticfiles %}
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>Reformed Academy</title>
-    <link rel="stylesheet" href="http://maxcdn.bootstrapcdn.com/bootstrap/3.2.0/css/bootstrap.min.css">
-    <link href='http://fonts.googleapis.com/css?family=Questrial|Muli|Open+Sans' rel='stylesheet' type='text/css'>
-
-    <!-- HTML5 Shim and Respond.js IE8 support of HTML5 elements and media queries -->
-    <!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
-    <!--[if lt IE 9]>
-      <script src="https://oss.maxcdn.com/html5shiv/3.7.2/html5shiv.min.js"></script>
-      <script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
-    <![endif]-->
-
-<!--     <link rel="stylesheet" href="http://maxcdn.bootstrapcdn.com/bootstrap/3.2.0/css/bootstrap-theme.min.css"> -->
-    <link href="http://maxcdn.bootstrapcdn.com/font-awesome/4.1.0/css/font-awesome.min.css" rel="stylesheet">
-    <link rel="stylesheet" href="{% static 'css/main.css' %}">
-
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.1/jquery.min.js"></script>
-    <script src="http://maxcdn.bootstrapcdn.com/bootstrap/3.2.0/js/bootstrap.min.js"></script>
-    <script>
-        $(document).mouseup(function (e) {
-            var container = $(".menu");
-
-            if (container.hasClass('open') && !container.is(e.target)
-                && container.has(e.target).length === 0) {
-                container.removeClass('open');
-            }
-        });
-        $(document).ready(function() {
-
-            $("#logo").click(function() {
-                $('.menu').toggleClass('open');
-            });
-        });
-    </script>
-  </head>
-  <body>
-    <nav class="menu">
-        <ul>
-            <li>
-                <a href="#"><i class="fa fa-home"></i>Home</a>
-            </li>
-            <li>
-                <a href="#"><i class="fa fa-graduation-cap"></i>Courses</a>
-            </li>
-        </ul>
-
-        <ul class="profile">
-            <li>
-                <a href="#"><i class="fa fa-tasks"></i>Progress</a>
-            </li>
-            <li>
-                <a href="#"><i class="fa fa-user"></i>Edit</a>
-            </li>
-            <li>
-                <a href="#"><i class="fa fa-lock"></i>Sign Out</a>
-            </li>
-        </ul>
-    </nav>
-    <div class="header">
-        <div class="logo">
-            <img id="logo" src="{% static 'images/lion.png' %}" alt=Logo">
-        </div>
-        <!--
-<div class="profile">
-            <img src="static/images/portrait.png" class="profile-image">
-        </div>
- -->
-    </div>
-    {% block content %}{% endblock %}
-  </body>
-</html>
