--- reformedacademy/templates/reformedacademy/email/activation.html
@@ -1,3 +1,0 @@
-Thank you for signing up for Reformed Academy! Please click on the link below to finish creating your account.
-
-{{ url }}
