--- reformedacademy/templates/reformedacademy/login.html
@@ -1,27 +1,0 @@
-{% extends 'reformedacademy/base_no_menu.html' %}
-{% load staticfiles %}
-
-{% block content %}
-    <div class="top-container centered">
-        <div class="inner">
-            <img src="{% static 'images/lion.png' %}" alt="Logo">
-            <h1>Log In</h1>
-            <div class="description">
-                Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.
-            </div>
-        </div>
-    </div>
-    <div class="main-container centered">
-        <div class="inner">
-            <form action="" method="POST" class="login-form">
-                {% csrf_token %}
-                {{ form }}
-                <input type="submit" value="Log In">
-            </form>
-            <div class="previous-account">
-                Don't have an account?
-                <a href="{% url 'signup' %}">Sign Up</a>
-            </div>
-        </div>
-    </div>
-{% endblock %}
