--- reformedacademy/templates/reformedacademy/subscribe.html
@@ -1,71 +1,0 @@
-{% load staticfiles %}
-
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>Reformed Academy</title>
-    <link rel="stylesheet" href="http://maxcdn.bootstrapcdn.com/bootstrap/3.2.0/css/bootstrap.min.css">
-    <link href='http://fonts.googleapis.com/css?family=Questrial|Open+Sans' rel='stylesheet' type='text/css'>
-
-    <!-- HTML5 Shim and Respond.js IE8 support of HTML5 elements and media queries -->
-    <!-- WARNING: Respond.js doesn't work if you view the page via file:// -->
-    <!--[if lt IE 9]>
-      <script src="https://oss.maxcdn.com/html5shiv/3.7.2/html5shiv.min.js"></script>
-      <script src="https://oss.maxcdn.com/respond/1.4.2/respond.min.js"></script>
-    <![endif]-->
-
-<!--     <link rel="stylesheet" href="http://maxcdn.bootstrapcdn.com/bootstrap/3.2.0/css/bootstrap-theme.min.css"> -->
-    <link href="http://maxcdn.bootstrapcdn.com/font-awesome/4.1.0/css/font-awesome.min.css" rel="stylesheet">
-    <link rel="stylesheet" href="{% static 'css/main.css' %}">
-
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.11.1/jquery.min.js"></script>
-    <script src="http://maxcdn.bootstrapcdn.com/bootstrap/3.2.0/js/bootstrap.min.js"></script>
-  </head>
-  <body>
-    <script>
-     (function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){
-     (i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),
-     m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)
-     })(window,document,'script','//www.google-analytics.com/analytics.js','ga');
-
-     ga('create', 'UA-79589-16', 'auto');
-     ga('send', 'pageview');
-
-    </script>
-    <div class="signup-top-container">
-        <div class="inner">
-            <img src="{% static 'images/lion.png' %}" alt="Logo">
-            <h1>Reformed Academy</h1>
-            <div class="description">
-                <div>Be the first to know more information on Reformed Academy by subscribing to our mailing list.</div>
-                <div>Reformed Academy is a project of <a href="http://reformedforum.org">Reformed Forum</a>, a 501(c)(3) organization. Learn more about Reformed Academy <a href="http://reformedforum.org/need-help-new-christian-education-initiative/?utm_content=buffer4137a&utm_medium=social&utm_source=twitter.com&utm_campaign=buffer">here</a> and <a href="https://medium.com/@ReformedAcademy/something-different-896fe9b02648">here</a>.</div>
-            </div>
-        </div>
-    </div>
-    <div class="signup-container">
-        <div class="inner">
-            <div class="signup-form">
-            <div id="mc_embed_signup">
-            <form action="//reformedforum.us2.list-manage.com/subscribe/post?u=14186c205d6f7604ade1f3155&amp;id=45bbbe36dd" method="post" id="mc-embedded-subscribe-form" name="mc-embedded-subscribe-form" class="validate" target="_blank" novalidate>
-                <label for="mce-EMAIL">Subscribe to our mailing list</label>
-                <input type="email" value="" name="EMAIL" class="email" id="mce-EMAIL" placeholder="email address" required>
-                <!-- real people should not fill this in and expect good things - do not remove this or risk form bot signups-->
-                <div style="position: absolute; left: -5000px;"><input type="text" name="b_f3d57ec181c2e66831b41ddff_15cdb0f60f" tabindex="-1" value=""></div>
-                <div class="terms-and-conditions">
-                    Please read our <a href="http://reformedforum.org/about/privacy/" target="_blank">Privacy Policy</a> for more information on how your email address will be protected.
-                </div>
-                <div class="clear"><input type="submit" value="Subscribe" name="subscribe" id="mc-embedded-subscribe" class="button"></div>
-            </form>
-            </div>
-            <div class="social">
-                <a href="http://twitter.com/ReformedAcademy"><i class="fa fa-twitter"></i></ad>
-            </div>
-        </div>
-    </div>
-  </body>
-</html>
-
-
