--- reformedacademy/templates/reformedacademy/welcome.html
@@ -1,13 +1,0 @@
-{% extends 'reformedacademy/base.html' %}
-{% load staticfiles %}
-
-{% block content %}
-    <div class="top-container">
-        <div class="inner">
-            <h1>Welcome</h1>
-            <div class="description">
-                Welcome to Reformed Academy!
-            </div>
-        </div>
-    </div>
-{% endblock %}
