+++ requirements.txt
@@ -1,6 +1,5 @@
 Django==1.6.5
 MySQL-python==1.2.5
-Pillow==2.5.1
 boto==2.29.1
 django-ipware==0.0.8
 django-storages==1.1.8
